# server/docs/swagger.yaml
definitions:
...
  CreateLoginRequest:
    properties:
      password:
        minLength: 8
        type: string
      role:
        enum:
        - bot
        - executive
        - tournament_director
        - secretary
        - treasurer
        - vice_president
        - president
        - webmaster
        type: string
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
...
  ErrorResponse:
    properties:
      code:
        description: Code is the HTTP status code for the error.
        type: integer
      message:
        description: Message is a more descriptive error message.
        type: string
      type:
        description: Type is the name of the HTTP status code.
        type: string
    type: object
...
  LinkedMemberInfo:
    properties:
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    type: object
  LoginWithMember:
    properties:
      linkedMember:
        $ref: '#/definitions/LinkedMemberInfo'
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
...
  UpdateLoginRequest:
    properties:
      password:
        minLength: 8
        type: string
      role:
        enum:
        - bot
        - executive
        - tournament_director
        - secretary
        - treasurer
        - vice_president
        - president
        - webmaster
        type: string
      status:
        enum:
        - active
        - pending_activation
        - disabled
        type: string
    type: object
...
info:
  contact:
    email: uwaterloopoker@gmail.com
    name: UWPSC Development Team
  description: This is the API for the UWPSC website.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: UWPSC API
  version: "1.0"
paths:
...
  /logins:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all logins with their linked member information
      parameters:
      - description: Search term to filter logins by username, role, or linked member
          name (case-insensitive)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/LoginWithMember'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List all logins
      tags:
      - Logins
    post:
      consumes:
      - application/json
      description: Create a new login with the provided credentials
      parameters:
      - description: Login details
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/CreateLoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/LoginWithMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new login
      tags:
      - Logins
  /logins/{username}:
    delete:
      consumes:
      - application/json
      description: Delete a login by username
      parameters:
      - description: Login username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete login by username
      tags:
      - Logins
    get:
      consumes:
      - application/json
      description: Retrieve a login by username with linked member information
      parameters:
      - description: Login username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginWithMember'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get login by username
      tags:
      - Logins
    patch:
      consumes:
      - application/json
      description: Update a login's password and/or role. At least one field must
        be provided.
      parameters:
      - description: Login username
        in: path
        name: username
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateLoginRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update a login
      tags:
      - Logins
...
swagger: "2.0"