# server/internal/controller/logins.go
package controller

import (
	apierrors "api/internal/errors"
	"api/internal/middleware"
	"api/internal/models"
	"api/internal/services"
	"api/internal/store"
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"
)

type loginsController struct {
	store store.Store
}

// NewLoginsController creates a new instance of loginsController
func NewLoginsController(st store.Store) Controller {
	return &loginsController{store: st}
}

func (c *loginsController) LoadRoutes(router *gin.RouterGroup) {
	logins := router.Group("logins", middleware.UseAuthentication(c.store))
	logins.GET("", middleware.UseAuthorization("login.list"), c.listLogins)
	logins.GET("/:username", middleware.UseAuthorization("login.get"), c.getLogin)
	logins.POST("", middleware.UseAuthorization("login.create"), c.createLogin)
	logins.DELETE("/:username", middleware.UseAuthorization("login.delete"), c.deleteLogin)
	logins.PATCH("/:username", middleware.UseAuthorization("login.edit"), c.updateLogin)
}

// listLogins handles listing all logins with linked member information
//
// @Summary List all logins
// @Description Retrieve a list of all logins with their linked member information
// @Tags Logins
// @Accept json
// @Produce json
// @Param search query string false "Search term to filter logins by username, role, or linked member name (case-insensitive)"
// @Success 200 {array} LoginWithMember
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /logins [get]
func (c *loginsController) listLogins(ctx *gin.Context) {
	pagination, err := models.ParsePagination(ctx)
	if err != nil {
		ctx.AbortWithStatusJSON(http.StatusBadRequest, apierrors.InvalidRequest(err.Error()))
		return
	}

	search := ctx.Query("search")

	logins, total, err := c.store.Logins().List(&pagination, search)
	if err != nil {
		ctx.AbortWithStatusJSON(
			http.StatusInternalServerError,
			apierrors.InternalServerError(err.Error()),
		)
		return
	}

	ctx.JSON(http.StatusOK, models.ListResponse[models.LoginWithMember]{
		Data:  logins,
		Total: total,
	})
}

// getLogin handles retrieving a single login by username
//
// @Summary Get login by username
// @Description Retrieve a login by username with linked member information
// @Tags Logins
// @Accept json
// @Produce json
// @Param username path string true "Login username"
// @Success 200 {object} LoginWithMember
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /logins/{username} [get]
func (c *loginsController) getLogin(ctx *gin.Context) {
	username := ctx.Param("username")
	if username == "" {
		ctx.AbortWithStatusJSON(
			http.StatusBadRequest,
			apierrors.InvalidRequest("username parameter is required"),
		)
		return
	}

	login, err := c.store.Logins().FindByUsernameWithMember(username)
	if err != nil {
		if errors.Is(err, store.ErrNotFound) {
			ctx.AbortWithStatusJSON(http.StatusNotFound, apierrors.NotFound("login not found"))
			return
		}
		ctx.AbortWithStatusJSON(
			http.StatusInternalServerError,
			apierrors.InternalServerError(err.Error()),
		)
		return
	}

	ctx.JSON(http.StatusOK, login)
}

// createLogin handles the creation of a new login
//
// @Summary Create a new login
// @Description Create a new login with the provided credentials
// @Tags Logins
// @Accept json
// @Produce json
// @Param login body CreateLoginRequest true "Login details"
// @Success 201 {object} LoginWithMember
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /logins [post]
func (c *loginsController) createLogin(ctx *gin.Context) {
	var req models.CreateLoginRequest
	if !BindJSON(ctx, &req) {
		return
	}

	svc := services.NewLoginService(c.store)
	err := svc.CreateLogin(req.Username, req.Password, req.Role)
	if err != nil {
		ctx.AbortWithStatusJSON(
			http.StatusInternalServerError,
			apierrors.InternalServerError(err.Error()),
		)
		return
	}

	// Return the created login (without password)
	login, err := c.store.Logins().FindByUsernameWithMember(req.Username)
	if err != nil {
		ctx.AbortWithStatusJSON(
			http.StatusInternalServerError,
			apierrors.InternalServerError(err.Error()),
		)
		return
	}

	ctx.JSON(http.StatusCreated, login)
}

// deleteLogin handles deleting a login by username
//
// @Summary Delete login by username
// @Description Delete a login by username
// @Tags Logins
// @Accept json
// @Produce json
// @Param username path string true "Login username"
// @Success 204 "No Content"
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /logins/{username} [delete]
func (c *loginsController) deleteLogin(ctx *gin.Context) {
	username := ctx.Param("username")
	if username == "" {
		ctx.AbortWithStatusJSON(
			http.StatusBadRequest,
			apierrors.InvalidRequest("username parameter is required"),
		)
		return
	}

	err := c.store.Logins().Delete(username)
	if err != nil {
		if errors.Is(err, store.ErrNotFound) {
			ctx.AbortWithStatusJSON(http.StatusNotFound, apierrors.NotFound("login not found"))
			return
		}
		ctx.AbortWithStatusJSON(
			http.StatusInternalServerError,
			apierrors.InternalServerError(err.Error()),
		)
		return
	}

	ctx.Status(http.StatusNoContent)
}

// updateLogin handles updating a login's password and/or role
//
// @Summary Update a login
// @Description Update a login's password and/or role. At least one field must be provided.
// @Tags Logins
// @Accept json
// @Produce json
// @Param username path string true "Login username"
// @Param request body UpdateLoginRequest true "Fields to update"
// @Success 204 "No Content"
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /logins/{username} [patch]
func (c *loginsController) updateLogin(ctx *gin.Context) {
	username := ctx.Param("username")
	if username == "" {
		ctx.AbortWithStatusJSON(
			http.StatusBadRequest,
			apierrors.InvalidRequest("username parameter is required"),
		)
		return
	}

	var req models.UpdateLoginRequest
	if !BindJSON(ctx, &req) {
		return
	}

	svc := services.NewLoginService(c.store)
	err := svc.UpdateLogin(username, req.Password, req.Role, req.Status)
	if err != nil {
		switch {
		case errors.Is(err, services.ErrLoginNotFound):
			ctx.AbortWithStatusJSON(http.StatusNotFound, apierrors.NotFound(err.Error()))
		case errors.Is(err, services.ErrUpdateLoginNoFields):
			ctx.AbortWithStatusJSON(http.StatusBadRequest, apierrors.InvalidRequest(err.Error()))
		default:
			ctx.AbortWithStatusJSON(
				http.StatusInternalServerError,
				apierrors.InternalServerError(err.Error()),
			)
		}
		return
	}

	ctx.Status(http.StatusNoContent)
}